Support wildcard for geofence name
Added support for a wildcard "*" in geofence trigger name.
This way it is possible to create a trigger that trigger for each
geofence.

In combination with the topic owntracks/+/+/event, you can create a trigger
that triggers for each user and for each geofence.

Also added the label 'fence' that contains the name of the geofence that is
entered or left.