Added the possibility for using autocomplete actions.
For autocomplte actions to work, we need to know what users are possible
and what geofences can be used. This is done automatically by creating
arrays of users, the state of the user and geofences that are entered or left.

Up to now, all gathered data was lost as soon as the app is restarted. So this
data is now being persisted. On loading of the app, two files are being read from
the filesystem that contains data of the user (name, fence the user was in, coordinates
of the user, battery percentage of the phone, timestamp of the last received message).
Next to the user data, also an array with the fences is read.

When the app is stopped, either by the user or by updating the app or rebooting of
Homey, the data is written to two files on the filesystem.

No history data is being saved.