Also use waypoint messages to detect if a new fence is added to a device.
Up to now, new geofence are automatically added when a user enters or leaves
the new geofence. Seems that as soon as a new fence is created in an app, it
sends out a waypoint message. This waypoint message also contains the name of
the fence that was added (if sharing is enabled).