Disable sending location data back to the client
At this time the Homey API adds status and result fields in the HTTTP
body. This might be nice for REST calls between the settings and main
business logic, but for handling third part interfaces this causes problems
because the added JSON fields can't be parsed by the owntracks phone apps.

So until there is a way to bypass the addition of the status and result fields,
i can't send the location data back to the phone apps.