Handle Publishing of Waypoints from phone app
From the owntracks phone apps you have the possibility to publish
all the waypoints / regions that are defined within the app.
This commit handles that messages and adds the received waypoints
to the geofence list with its name, coordinates and radius.

Also the settings page is adjusted to display this data.