Detect devicename and add to the userdata
When a MQTT message arrives on a topic, the devicename is part of the
topic. Take the devicename and put it to the user data for later use.

We are only going to handle ONE devicename per user. Multiple devices
per user does not have a usecase as you can't really be sure then where
the user actually is. The device name is needed to push messages to the
device (like a list of waypoints).