Add geofence pushing to owntracks clients
When using MQTT, it is possible to push commands to individual owntracks
phone clients. This commit implements creating and pushing a command
message with all the geofences that the Homey client has configured.
This requires the owntracks phone clients to have the "receive remote
commands" option set to enbaled.

Triggering the upload can be done for a device by pressing the '^'
button behind the user information line in the current user data tab.