owntracks: Make keepalive MQTT Client connect settings configurable.
Some reports came in that the connection to some brokers would be lost 
(resulting in a notification) and then reconnect again. By default the 
keepalive was set to 10 seconds. That would mean that every 10 seconds 
the MQTT client in owntracks would send a keepalive signal. If that 
signal was not received within 15 seconds, the broker would disconnect. 
However it seems that the broker receives a keepalive signal every 10 
seconds. The only explenation could be that sometimes the keepalive is 
not sent or lost.

This commit creates an extra setting in the settings page to enter the 
keepalive in seconds and that is used upon creating the connection with 
the broker.

When the keepalive setting is not entered, a default of 60 seconds is 
used.